requirements.txt
ctransformers==0.2.27
sentence-transformers==3.0.1
pinecone-client==4.1.1
langchain==0.2.6
flask==3.0.3
pypdf==4.2.0
python-dotenv==1.0.1
langchain-huggingface==0.0.3
-e .